+++ python/tests/test_serialization.py
@@ -2,6 +2,7 @@ import io
 import unittest
 
 import gtirb.serialization
+from gtirb.serialization import Variant
 
 
 class TestSerialization(unittest.TestCase):
@@ -154,6 +155,55 @@ class TestSerialization(unittest.TestCase):
             bstream.getvalue(), "UUID", ir.get_by_uuid
         )
         self.assertEqual(result, b3.uuid)
+
+    def test_variant_codec(self):
+        serializer = gtirb.serialization.Serialization()
+        ostream = io.BytesIO()
+        variant = Variant(2, {5: ["a", "b"], 15: ["cc", "ddd"]})
+        serializer.encode(
+            ostream,
+            variant,
+            "variant<string,int64_t,mapping<int64_t,sequence<string>>>",
+        )
+        raw_bytes = ostream.getvalue()
+        var_val = serializer.decode(
+            raw_bytes,
+            "variant<string,int64_t,mapping<int64_t,sequence<string>>>",
+        )
+        self.assertEqual(var_val, variant)
+
+        ostream = io.BytesIO()
+        variant = Variant(1, 10)
+        serializer.encode(
+            ostream, variant, "variant<string,int64_t,string>",
+        )
+        raw_bytes = ostream.getvalue()
+        var_val = serializer.decode(
+            raw_bytes, "variant<string,int64_t,string>"
+        )
+        self.assertEqual(var_val, variant)
+
+        ostream = io.BytesIO()
+        variant = Variant(0, "zzzz")
+        serializer.encode(
+            ostream, variant, "variant<string,int64_t,string>",
+        )
+        raw_bytes = ostream.getvalue()
+        var_val = serializer.decode(
+            raw_bytes, "variant<string,int64_t,string>"
+        )
+        self.assertEqual(var_val, variant)
+
+        ostream = io.BytesIO()
+        mapping = {"aa": Variant(0, 5), "bbb": Variant(1, "ccccc")}
+        serializer.encode(
+            ostream, mapping, "mapping<string,variant<int64_t,string>>",
+        )
+        raw_bytes = ostream.getvalue()
+        mapping_val = serializer.decode(
+            raw_bytes, "mapping<string,variant<int64_t,string>>"
+        )
+        self.assertEqual(mapping_val, mapping)
 
 
 if __name__ == "__main__":